backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
supabase==2.3.4
pytest==8.0.2
black==24.2.0
flake8==7.0.0
isort==5.13.2
gunicorn==21.2.0
python-jose==3.3.0
requests==2.31.0
Pillow>=9.5.0
python-magic==0.4.27 